test.1a01/requirements.txt
Django >= 2.2.10
djangorestframework
